+++ tests/Functional/TaxonomiesTest.php
@@ -18,6 +18,7 @@ use Illuminate\Support\Collection;
 use PHPUnit\Framework\TestCase;
 use VaniloCloud\ApiClient;
 use VaniloCloud\Models\Taxonomy;
+use VaniloCloud\Stores\ApcTokenStore;
 
 class TaxonomiesTest extends TestCase
 {